GFETCH-119 organizes classes into packages
