+++ app/src/main/java/co/gobd/gofetch/adapter/SupportedOrderAdapter.java
@@ -16,7 +16,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import co.gobd.gofetch.R;
-import co.gobd.gofetch.listener.OnItemClickListener;
+import co.gobd.gofetch.ui.view.OnItemClickListener;
 import co.gobd.gofetch.model.SupportedOrder;
 import co.gobd.gofetch.service.order.OrderService;
 import co.gobd.gofetch.service.order.SupportedOrderCallback;
