--- app/src/main/java/co/gobd/gofetch/config/BuildConfiguration.java
@@ -1,9 +1,0 @@
-package co.gobd.gofetch.config;
-
-/**
- * Created by tonmoy on 01-Feb-16.
- */
-public class BuildConfiguration {
-
-    public static boolean IS_MOCK_DATA = true;
-}
