--- app/src/main/java/co/gobd/gofetch/config/SignalRConfig.java
@@ -1,10 +1,0 @@
-package co.gobd.gofetch.config;
-
-/**
- * Created by tonmoy on 04-Feb-16.
- */
-public class SignalRConfig {
-
-    public final static String URL = "http://gobdodin.cloudapp.net:1337/realtime/hubs";
-    public final static String HUB_NAME = "GoTransitBroadcastHub";
-}
