+++ app/src/main/java/co/gobd/gofetch/di/component/AppComponent.java
@@ -2,7 +2,7 @@ package co.gobd.gofetch.di.component;
 
 import javax.inject.Singleton;
 
-import co.gobd.gofetch.activity.SupportedOrderActivity;
+import co.gobd.gofetch.ui.activity.SupportedOrderActivity;
 import co.gobd.gofetch.di.module.ApiModule;
 import co.gobd.gofetch.di.module.AppModule;
 import co.gobd.gofetch.di.module.ServiceModule;
