+++ app/src/main/java/co/gobd/gofetch/ui/activity/RideActivity.java
@@ -1,12 +1,12 @@
-package co.gobd.gofetch.activity;
+package co.gobd.gofetch.ui.activity;
 
 import android.content.res.Configuration;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 import co.gobd.gofetch.R;
-import co.gobd.gofetch.callback.RideFragmentCallback;
-import co.gobd.gofetch.fragment.RoutePlanFragment;
+import co.gobd.gofetch.ui.fragment.RideFragmentCallback;
+import co.gobd.gofetch.ui.fragment.RoutePlanFragment;
 
 public class RideActivity extends AppCompatActivity implements RideFragmentCallback {
