+++ app/src/main/java/co/gobd/gofetch/utility/LocationType.java
@@ -1,9 +1,9 @@
-package co.gobd.gofetch.enums;
+package co.gobd.gofetch.utility;
 
 /**
  * Created by tonmoy on 10-Mar-16.
  */
-public enum LocationTypeEnum {
+public enum LocationType {
     STARTING_POINT,
     DESTINATION_POINT
