+++ app/src/main/res/layout/activity_ride.xml
@@ -5,6 +5,6 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
 
-    tools:context="co.gobd.gofetch.activity.RideActivity">
+    tools:context=".ui.activity.RideActivity">
 
 </RelativeLayout>
