+++ app/src/main/res/layout/activity_supported_order.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".activity.SupportedOrderActivity">
+    tools:context=".ui.activity.SupportedOrderActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
